Getting cards for locations to append to page, while adding css. Additionally  moving locationCard, locationPoint, and simpleMap to components file.
