+++ src/components/LocationPoint.js
